# api/swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: freekik
  description: 'Api docs freekik'
host: localhost:3000
basePath: /api
schemes:
- http
consumes:
- application/json
produces:
- application/json
paths:
...
  /post:
    get:
      description: 'TODO: Add Description'
      summary: get my posts
      tags:
      - posts
      operationId: PostGet
      produces:
      - application/json
      parameters:
      - name: limit
        in: query
        required: false
        type: string
        description: ''
      - name: skip
        in: query
        required: false
        type: string
        description: ''
      - name: x-access-token
        in: header
        required: true
        type: string
        description: ''
      responses:
        200:
          description: ''
          schema:
            $ref: '#/definitions/GetMyPosts'
          examples:
            application/json:
              success: true
              response:
              - deactivated: false
                taggedUsers:
                - 5ab52232d28be4ca7d30a324
                _id: 5ab53032633d761b496bc34c
                title: Javascript, the good parts
                body: Lorem Ipsum Dolor Sit Amet
                picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
                creator: 5ab52232d28be4ca7d30a324
                createdDate: 2018-03-23T16:49:54.4800000Z
                updatedDate: 2018-03-24T15:22:20.7430000Z
                __v: 0
              - deactivated: false
                taggedUsers:
                - 5ab52232d28be4ca7d30a324
                _id: 5ab537c4d8b4081d068088dd
                title: Javascript, the goods parts
                body: Lorem Ipsum Dolor Sit Amet
                picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
                creator: 5ab52232d28be4ca7d30a324
                createdDate: 2018-03-23T17:22:12.1840000Z
                updatedDate: 2018-03-23T17:22:12.1840000Z
                __v: 0
              - deactivated: false
                taggedUsers:
                - 5ab52232d28be4ca7d30a324
                _id: 5ab537e2fa50d61d2ae6917e
                title: Javascript, the goods parts
                body: Lorem Ipsum Dolor Sit Amet
                picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
                creator: 5ab52232d28be4ca7d30a324
                createdDate: 2018-03-23T17:22:42.9740000Z
                updatedDate: 2018-03-23T17:22:42.9740000Z
                __v: 0
              - deactivated: false
                taggedUsers:
                - 5ab52232d28be4ca7d30a324
                - 5ab52232d28be4ca7d30a325
                _id: 5ab538d7fa50d61d2ae6917f
                title: Javascript, the goods parts
                body: Lorem Ipsum Dolor Sit Amet
                picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
                creator: 5ab52232d28be4ca7d30a324
                createdDate: 2018-03-23T17:26:47.0150000Z
                updatedDate: 2018-03-23T17:26:47.0150000Z
                __v: 0
              - deactivated: false
                taggedUsers:
                - 5ab52232d28be4ca7d30a324
                - 5ab52232d28be4ca7d30a325
                _id: 5ab6516e1ab22944e026ba54
                title: Javascript, the goods parts
                body: Lorem Ipsum Dolor Sit Amet
                picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
                creator: 5ab52232d28be4ca7d30a324
                createdDate: 2018-03-24T13:23:58.0560000Z
                updatedDate: 2018-03-24T13:23:58.0560000Z
                __v: 0
              - deactivated: false
                taggedUsers:
                - 5ab52232d28be4ca7d30a324
                - 5ab52232d28be4ca7d30a325
                _id: 5ab651f8cef4b94539fa0508
                title: Javascript, the goods parts
                body: Lorem Ipsum Dolor Sit Amet
                picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
                creator: 5ab52232d28be4ca7d30a324
                createdDate: 2018-03-24T13:26:16.4700000Z
                updatedDate: 2018-03-24T13:26:16.4700000Z
                __v: 0
              - deactivated: false
                taggedUsers:
                - 5ab52232d28be4ca7d30a324
                - 5ab52232d28be4ca7d30a325
                _id: 5ab65242b76201454a6c9289
                title: Javascript, the bad parts
                body: Lorem Ipsum Dolor Sit Amet
                picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
                creator: 5ab52232d28be4ca7d30a324
                createdDate: 2018-03-24T13:27:30.3940000Z
                updatedDate: 2018-03-24T13:27:30.3940000Z
                __v: 0
              - deactivated: false
                taggedUsers:
                - 5ab52232d28be4ca7d30a324
                _id: 5ab66c4d03f5314f4123906e
                title: Javascript, the good parts
                body: Lorem Ipsum Dolor Sit Amet
                picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
                creator: 5ab52232d28be4ca7d30a324
                createdDate: 2018-03-24T15:18:37.8570000Z
                updatedDate: 2018-03-24T15:18:37.8570000Z
                __v: 0
              - deactivated: false
                taggedUsers:
                - 5ab52232d28be4ca7d30a324
                _id: 5ab66d3c03f5314f4123906f
                title: Javascript, the good parts
                body: Lorem Ipsum Dolor Sit Amet
                picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
                creator: 5ab52232d28be4ca7d30a324
                createdDate: 2018-03-24T15:22:36.4780000Z
                updatedDate: 2018-03-24T15:22:36.4780000Z
                __v: 0
      security: []
      x-unitTests:
      - request:
          method: GET
          uri: /post?limit={{limit}}&skip={{skip}}
          headers:
            x-access-token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI1YWI1MjIzMmQyOGJlNGNhN2QzMGEzMjQiLCJpYXQiOjE1MjE4OTU2Mjh9.GOEJXmpIA08J3vhIly2GOsqZsfEXOcMC3VZPO3Fdg_U
        expectedResponse:
          x-allowExtraHeaders: true
          x-bodyMatchMode: RAW
          x-arrayOrderedMatching: false
          x-arrayCheckCount: false
          x-matchResponseSchema: true
          headers:
            Access-Control-Allow-Origin: '*'
            Connection: keep-alive
            Content-Length: 3516
            Content-Type: application/json; charset=utf-8
            Date: Sat, 24 Mar 2018 15:23:46 GMT
            ETag: W/"dbc-c8+TnxhS7qS5m/cj5kCzG9k0UVw"
            X-Powered-By: Express
          body: '{"success":true,"response":[{"deactivated":false,"taggedUsers":["5ab52232d28be4ca7d30a324"],"_id":"5ab53032633d761b496bc34c","title":"Javascript, the good parts","body":"Lorem Ipsum Dolor Sit Amet","picture":"https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg","creator":"5ab52232d28be4ca7d30a324","createdDate":"2018-03-23T16:49:54.48Z","updatedDate":"2018-03-24T15:22:20.743Z","__v":0},{"deactivated":false,"taggedUsers":["5ab52232d28be4ca7d30a324"],"_id":"5ab537c4d8b4081d068088dd","title":"Javascript, the goods parts","body":"Lorem Ipsum Dolor Sit Amet","picture":"https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg","creator":"5ab52232d28be4ca7d30a324","createdDate":"2018-03-23T17:22:12.184Z","updatedDate":"2018-03-23T17:22:12.184Z","__v":0},{"deactivated":false,"taggedUsers":["5ab52232d28be4ca7d30a324"],"_id":"5ab537e2fa50d61d2ae6917e","title":"Javascript, the goods parts","body":"Lorem Ipsum Dolor Sit Amet","picture":"https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg","creator":"5ab52232d28be4ca7d30a324","createdDate":"2018-03-23T17:22:42.974Z","updatedDate":"2018-03-23T17:22:42.974Z","__v":0},{"deactivated":false,"taggedUsers":["5ab52232d28be4ca7d30a324","5ab52232d28be4ca7d30a325"],"_id":"5ab538d7fa50d61d2ae6917f","title":"Javascript, the goods parts","body":"Lorem Ipsum Dolor Sit Amet","picture":"https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg","creator":"5ab52232d28be4ca7d30a324","createdDate":"2018-03-23T17:26:47.015Z","updatedDate":"2018-03-23T17:26:47.015Z","__v":0},{"deactivated":false,"taggedUsers":["5ab52232d28be4ca7d30a324","5ab52232d28be4ca7d30a325"],"_id":"5ab6516e1ab22944e026ba54","title":"Javascript, the goods parts","body":"Lorem Ipsum Dolor Sit Amet","picture":"https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg","creator":"5ab52232d28be4ca7d30a324","createdDate":"2018-03-24T13:23:58.056Z","updatedDate":"2018-03-24T13:23:58.056Z","__v":0},{"deactivated":false,"taggedUsers":["5ab52232d28be4ca7d30a324","5ab52232d28be4ca7d30a325"],"_id":"5ab651f8cef4b94539fa0508","title":"Javascript, the goods parts","body":"Lorem Ipsum Dolor Sit Amet","picture":"https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg","creator":"5ab52232d28be4ca7d30a324","createdDate":"2018-03-24T13:26:16.47Z","updatedDate":"2018-03-24T13:26:16.47Z","__v":0},{"deactivated":false,"taggedUsers":["5ab52232d28be4ca7d30a324","5ab52232d28be4ca7d30a325"],"_id":"5ab65242b76201454a6c9289","title":"Javascript, the bad parts","body":"Lorem Ipsum Dolor Sit Amet","picture":"https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg","creator":"5ab52232d28be4ca7d30a324","createdDate":"2018-03-24T13:27:30.394Z","updatedDate":"2018-03-24T13:27:30.394Z","__v":0},{"deactivated":false,"taggedUsers":["5ab52232d28be4ca7d30a324"],"_id":"5ab66c4d03f5314f4123906e","title":"Javascript, the good parts","body":"Lorem Ipsum Dolor Sit Amet","picture":"https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg","creator":"5ab52232d28be4ca7d30a324","createdDate":"2018-03-24T15:18:37.857Z","updatedDate":"2018-03-24T15:18:37.857Z","__v":0},{"deactivated":false,"taggedUsers":["5ab52232d28be4ca7d30a324"],"_id":"5ab66d3c03f5314f4123906f","title":"Javascript, the good parts","body":"Lorem Ipsum Dolor Sit Amet","picture":"https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg","creator":"5ab52232d28be4ca7d30a324","createdDate":"2018-03-24T15:22:36.478Z","updatedDate":"2018-03-24T15:22:36.478Z","__v":0}]}'
        x-testShouldPass: true
        x-testEnabled: true
        x-testName: get my posts
        x-testDescription: 'TODO: Add Description'
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
    post:
      description: 'TODO: Add Description'
      summary: create post
      tags:
      - posts
      operationId: PostPost
      produces:
      - application/json
      parameters:
      - name: Body
        in: body
        required: true
        description: ''
        schema:
          $ref: '#/definitions/CreatePostrequest'
      - name: x-access-token
        in: header
        required: true
        type: string
        description: ''
      responses:
        200:
          description: ''
          schema:
            $ref: '#/definitions/CreatePost'
          examples:
            application/json:
              success: true
              response:
                deactivated: false
                taggedUsers:
                - 5ab52232d28be4ca7d30a324
                _id: 5ab66d3c03f5314f4123906f
                title: Javascript, the good parts
                body: Lorem Ipsum Dolor Sit Amet
                picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
                creator: 5ab52232d28be4ca7d30a324
                createdDate: 2018-03-24T15:22:36.4780000Z
                updatedDate: 2018-03-24T15:22:36.4780000Z
                __v: 0
      security: []
      x-unitTests:
      - request:
          method: POST
          uri: /post
          headers:
            Content-Type: application/json
            x-access-token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI1YWI1MjIzMmQyOGJlNGNhN2QzMGEzMjQiLCJpYXQiOjE1MjE4OTU2Mjh9.GOEJXmpIA08J3vhIly2GOsqZsfEXOcMC3VZPO3Fdg_U
          body: '{"title":"Javascript, the good parts","body":"Lorem Ipsum Dolor Sit Amet","deactivated":false,"picture":"https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg","taggedUsers":["5ab52232d28be4ca7d30a324"]}'
        expectedResponse:
          x-allowExtraHeaders: true
          x-bodyMatchMode: RAW
          x-arrayOrderedMatching: false
          x-arrayCheckCount: false
          x-matchResponseSchema: true
          headers:
            Access-Control-Allow-Origin: '*'
            Connection: keep-alive
            Content-Length: 402
            Content-Type: application/json; charset=utf-8
            Date: Sat, 24 Mar 2018 15:22:36 GMT
            ETag: W/"192-pb/WuRhK+MOGq31AfzUODZxvOtc"
            X-Powered-By: Express
          body: '{"success":true,"response":{"deactivated":false,"taggedUsers":["5ab52232d28be4ca7d30a324"],"_id":"5ab66d3c03f5314f4123906f","title":"Javascript, the good parts","body":"Lorem Ipsum Dolor Sit Amet","picture":"https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg","creator":"5ab52232d28be4ca7d30a324","createdDate":"2018-03-24T15:22:36.478Z","updatedDate":"2018-03-24T15:22:36.478Z","__v":0}}'
        x-testShouldPass: true
        x-testEnabled: true
        x-testName: create post
        x-testDescription: 'TODO: Add Description'
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
...
definitions:
...
  GetMyPosts:
    title: get my posts
    example:
      success: true
      response:
      - deactivated: false
        taggedUsers:
        - 5ab52232d28be4ca7d30a324
        _id: 5ab53032633d761b496bc34c
        title: Javascript, the good parts
        body: Lorem Ipsum Dolor Sit Amet
        picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
        creator: 5ab52232d28be4ca7d30a324
        createdDate: 2018-03-23T16:49:54.4800000Z
        updatedDate: 2018-03-24T15:22:20.7430000Z
        __v: 0
      - deactivated: false
        taggedUsers:
        - 5ab52232d28be4ca7d30a324
        _id: 5ab537c4d8b4081d068088dd
        title: Javascript, the goods parts
        body: Lorem Ipsum Dolor Sit Amet
        picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
        creator: 5ab52232d28be4ca7d30a324
        createdDate: 2018-03-23T17:22:12.1840000Z
        updatedDate: 2018-03-23T17:22:12.1840000Z
        __v: 0
      - deactivated: false
        taggedUsers:
        - 5ab52232d28be4ca7d30a324
        _id: 5ab537e2fa50d61d2ae6917e
        title: Javascript, the goods parts
        body: Lorem Ipsum Dolor Sit Amet
        picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
        creator: 5ab52232d28be4ca7d30a324
        createdDate: 2018-03-23T17:22:42.9740000Z
        updatedDate: 2018-03-23T17:22:42.9740000Z
        __v: 0
      - deactivated: false
        taggedUsers:
        - 5ab52232d28be4ca7d30a324
        - 5ab52232d28be4ca7d30a325
        _id: 5ab538d7fa50d61d2ae6917f
        title: Javascript, the goods parts
        body: Lorem Ipsum Dolor Sit Amet
        picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
        creator: 5ab52232d28be4ca7d30a324
        createdDate: 2018-03-23T17:26:47.0150000Z
        updatedDate: 2018-03-23T17:26:47.0150000Z
        __v: 0
      - deactivated: false
        taggedUsers:
        - 5ab52232d28be4ca7d30a324
        - 5ab52232d28be4ca7d30a325
        _id: 5ab6516e1ab22944e026ba54
        title: Javascript, the goods parts
        body: Lorem Ipsum Dolor Sit Amet
        picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
        creator: 5ab52232d28be4ca7d30a324
        createdDate: 2018-03-24T13:23:58.0560000Z
        updatedDate: 2018-03-24T13:23:58.0560000Z
        __v: 0
      - deactivated: false
        taggedUsers:
        - 5ab52232d28be4ca7d30a324
        - 5ab52232d28be4ca7d30a325
        _id: 5ab651f8cef4b94539fa0508
        title: Javascript, the goods parts
        body: Lorem Ipsum Dolor Sit Amet
        picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
        creator: 5ab52232d28be4ca7d30a324
        createdDate: 2018-03-24T13:26:16.4700000Z
        updatedDate: 2018-03-24T13:26:16.4700000Z
        __v: 0
      - deactivated: false
        taggedUsers:
        - 5ab52232d28be4ca7d30a324
        - 5ab52232d28be4ca7d30a325
        _id: 5ab65242b76201454a6c9289
        title: Javascript, the bad parts
        body: Lorem Ipsum Dolor Sit Amet
        picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
        creator: 5ab52232d28be4ca7d30a324
        createdDate: 2018-03-24T13:27:30.3940000Z
        updatedDate: 2018-03-24T13:27:30.3940000Z
        __v: 0
      - deactivated: false
        taggedUsers:
        - 5ab52232d28be4ca7d30a324
        _id: 5ab66c4d03f5314f4123906e
        title: Javascript, the good parts
        body: Lorem Ipsum Dolor Sit Amet
        picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
        creator: 5ab52232d28be4ca7d30a324
        createdDate: 2018-03-24T15:18:37.8570000Z
        updatedDate: 2018-03-24T15:18:37.8570000Z
        __v: 0
      - deactivated: false
        taggedUsers:
        - 5ab52232d28be4ca7d30a324
        _id: 5ab66d3c03f5314f4123906f
        title: Javascript, the good parts
        body: Lorem Ipsum Dolor Sit Amet
        picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
        creator: 5ab52232d28be4ca7d30a324
        createdDate: 2018-03-24T15:22:36.4780000Z
        updatedDate: 2018-03-24T15:22:36.4780000Z
        __v: 0
    type: object
    properties:
      success:
        description: ''
        example: true
        type: boolean
      response:
        description: ''
        example:
        - deactivated: false
          taggedUsers:
          - 5ab52232d28be4ca7d30a324
          _id: 5ab53032633d761b496bc34c
          title: Javascript, the good parts
          body: Lorem Ipsum Dolor Sit Amet
          picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
          creator: 5ab52232d28be4ca7d30a324
          createdDate: 2018-03-23T16:49:54.4800000Z
          updatedDate: 2018-03-24T15:22:20.7430000Z
          __v: 0
        - deactivated: false
          taggedUsers:
          - 5ab52232d28be4ca7d30a324
          _id: 5ab537c4d8b4081d068088dd
          title: Javascript, the goods parts
          body: Lorem Ipsum Dolor Sit Amet
          picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
          creator: 5ab52232d28be4ca7d30a324
          createdDate: 2018-03-23T17:22:12.1840000Z
          updatedDate: 2018-03-23T17:22:12.1840000Z
          __v: 0
        - deactivated: false
          taggedUsers:
          - 5ab52232d28be4ca7d30a324
          _id: 5ab537e2fa50d61d2ae6917e
          title: Javascript, the goods parts
          body: Lorem Ipsum Dolor Sit Amet
          picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
          creator: 5ab52232d28be4ca7d30a324
          createdDate: 2018-03-23T17:22:42.9740000Z
          updatedDate: 2018-03-23T17:22:42.9740000Z
          __v: 0
        - deactivated: false
          taggedUsers:
          - 5ab52232d28be4ca7d30a324
          - 5ab52232d28be4ca7d30a325
          _id: 5ab538d7fa50d61d2ae6917f
          title: Javascript, the goods parts
          body: Lorem Ipsum Dolor Sit Amet
          picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
          creator: 5ab52232d28be4ca7d30a324
          createdDate: 2018-03-23T17:26:47.0150000Z
          updatedDate: 2018-03-23T17:26:47.0150000Z
          __v: 0
        - deactivated: false
          taggedUsers:
          - 5ab52232d28be4ca7d30a324
          - 5ab52232d28be4ca7d30a325
          _id: 5ab6516e1ab22944e026ba54
          title: Javascript, the goods parts
          body: Lorem Ipsum Dolor Sit Amet
          picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
          creator: 5ab52232d28be4ca7d30a324
          createdDate: 2018-03-24T13:23:58.0560000Z
          updatedDate: 2018-03-24T13:23:58.0560000Z
          __v: 0
        - deactivated: false
          taggedUsers:
          - 5ab52232d28be4ca7d30a324
          - 5ab52232d28be4ca7d30a325
          _id: 5ab651f8cef4b94539fa0508
          title: Javascript, the goods parts
          body: Lorem Ipsum Dolor Sit Amet
          picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
          creator: 5ab52232d28be4ca7d30a324
          createdDate: 2018-03-24T13:26:16.4700000Z
          updatedDate: 2018-03-24T13:26:16.4700000Z
          __v: 0
        - deactivated: false
          taggedUsers:
          - 5ab52232d28be4ca7d30a324
          - 5ab52232d28be4ca7d30a325
          _id: 5ab65242b76201454a6c9289
          title: Javascript, the bad parts
          body: Lorem Ipsum Dolor Sit Amet
          picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
          creator: 5ab52232d28be4ca7d30a324
          createdDate: 2018-03-24T13:27:30.3940000Z
          updatedDate: 2018-03-24T13:27:30.3940000Z
          __v: 0
        - deactivated: false
          taggedUsers:
          - 5ab52232d28be4ca7d30a324
          _id: 5ab66c4d03f5314f4123906e
          title: Javascript, the good parts
          body: Lorem Ipsum Dolor Sit Amet
          picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
          creator: 5ab52232d28be4ca7d30a324
          createdDate: 2018-03-24T15:18:37.8570000Z
          updatedDate: 2018-03-24T15:18:37.8570000Z
          __v: 0
        - deactivated: false
          taggedUsers:
          - 5ab52232d28be4ca7d30a324
          _id: 5ab66d3c03f5314f4123906f
          title: Javascript, the good parts
          body: Lorem Ipsum Dolor Sit Amet
          picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
          creator: 5ab52232d28be4ca7d30a324
          createdDate: 2018-03-24T15:22:36.4780000Z
          updatedDate: 2018-03-24T15:22:36.4780000Z
          __v: 0
        type: array
        items:
          $ref: '#/definitions/Response'
    required:
    - success
    - response
  Response:
    title: Response
    example:
      deactivated: false
      taggedUsers:
      - 5ab52232d28be4ca7d30a324
      _id: 5ab53032633d761b496bc34c
      title: Javascript, the good parts
      body: Lorem Ipsum Dolor Sit Amet
      picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
      creator: 5ab52232d28be4ca7d30a324
      createdDate: 2018-03-23T16:49:54.4800000Z
      updatedDate: 2018-03-24T15:22:20.7430000Z
      __v: 0
    type: object
    properties:
      deactivated:
        description: ''
        example: false
        type: boolean
      taggedUsers:
        description: ''
        example:
        - 5ab52232d28be4ca7d30a324
        type: array
        items:
          type: string
      _id:
        description: ''
        example: 5ab53032633d761b496bc34c
        type: string
      title:
        description: ''
        example: Javascript, the good parts
        type: string
      body:
        description: ''
        example: Lorem Ipsum Dolor Sit Amet
        type: string
      picture:
        description: ''
        example: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
        type: string
      creator:
        description: ''
        example: 5ab52232d28be4ca7d30a324
        type: string
      createdDate:
        description: ''
        example: 3/23/2018 4:49:54 PM
        type: string
      updatedDate:
        description: ''
        example: 3/24/2018 3:22:20 PM
        type: string
      __v:
        description: ''
        example: 0
        type: integer
        format: int32
    required:
    - deactivated
    - taggedUsers
    - _id
    - title
    - body
    - picture
    - creator
    - createdDate
    - updatedDate
    - __v
...
  CreatePostrequest:
    title: create postRequest
    example:
      title: Javascript, the good parts
      body: Lorem Ipsum Dolor Sit Amet
      deactivated: false
      picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
      taggedUsers:
      - 5ab52232d28be4ca7d30a324
    type: object
    properties:
      title:
        description: ''
        example: Javascript, the good parts
        type: string
      body:
        description: ''
        example: Lorem Ipsum Dolor Sit Amet
        type: string
      deactivated:
        description: ''
        example: false
        type: boolean
      picture:
        description: ''
        example: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
        type: string
      taggedUsers:
        description: ''
        example:
        - 5ab52232d28be4ca7d30a324
        type: array
        items:
          type: string
    required:
    - title
    - body
    - deactivated
    - picture
    - taggedUsers
  CreatePost:
    title: create post
    example:
      success: true
      response:
        deactivated: false
        taggedUsers:
        - 5ab52232d28be4ca7d30a324
        _id: 5ab66d3c03f5314f4123906f
        title: Javascript, the good parts
        body: Lorem Ipsum Dolor Sit Amet
        picture: https://images-na.ssl-images-amazon.com/images/I/81kqrwS1nNL.jpg
        creator: 5ab52232d28be4ca7d30a324
        createdDate: 2018-03-24T15:22:36.4780000Z
        updatedDate: 2018-03-24T15:22:36.4780000Z
        __v: 0
    type: object
    properties:
      success:
        description: ''
        example: true
        type: boolean
      response:
        $ref: '#/definitions/Response'
    required:
    - success
    - response
...